+++ qt_projet_2A/qt_projet_2A.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 3.0.1, 2016-05-26T15:57:58. -->
+<!-- Written by QtCreator 3.0.1, 2016-05-27T12:34:11. -->
 <qtcreator>
  <data>
   <variable>ProjectExplorer.Project.ActiveTarget</variable>
